Updated departmentControllerUnitTesting everything works so far

+++ basicCrud/src/main/java/com/example/task/basic/crud/basicCrud/service/impl/DepartmentImpl.java
@@ -2,6 +2,7 @@ package com.example.task.basic.crud.basicCrud.service.impl;
 
 import com.example.task.basic.crud.basicCrud.model.Department;
 import com.example.task.basic.crud.basicCrud.model.dto.DepartmentDTO;
+import com.example.task.basic.crud.basicCrud.model.exceptions.BadRequestException;
 import com.example.task.basic.crud.basicCrud.model.exceptions.DepartmentNotFoundException;
 import com.example.task.basic.crud.basicCrud.model.exceptions.InvalidDepartmentDtoException;
 import com.example.task.basic.crud.basicCrud.model.exceptions.InvalidDepartmentIdException;
@@ -26,9 +27,9 @@ public class DepartmentImpl implements DepartmentService {
     @Override
     public DepartmentDTO save(DepartmentDTO departmentDTO) {
         if(departmentDTO.getId() != null) {
-            throw new IllegalArgumentException("You cannot send id when creating a department");
+            throw new BadRequestException("You cannot send id when creating a department");
         }
-        if(departmentDTO.getName()!=null && !departmentDTO.getName().isEmpty()){
+        if(departmentDTO.getName() != null && !departmentDTO.getName().isEmpty()){
             Department department = Department.builder()
                     .name(departmentDTO.getName())
                     .build();
@@ -36,7 +37,7 @@ public class DepartmentImpl implements DepartmentService {
             departmentRepository.save(department);
             return DepartmentMapper.INSTANCE.departmentToDepartmentDTO(department);
         } else {
-            throw new IllegalArgumentException("Invalid department name");
+            throw new BadRequestException("Invalid department name");
         }
     }
 
@@ -54,12 +55,20 @@ public class DepartmentImpl implements DepartmentService {
 
     @Override
     public DepartmentDTO getDepartmentById(String id) {
+        if(isValidUUID(id)) {
+            Department department = departmentRepository.findById(UUID.fromString(id)).orElseThrow(InvalidDepartmentIdException::new);
+            return DepartmentMapper.INSTANCE.departmentToDepartmentDTO(department);
+        } else  {
+            throw new BadRequestException("Invalid id");
+        }
+    }
+
+    private static boolean isValidUUID(String id) {
         try {
             UUID uuid = UUID.fromString(id);
-            Department department = departmentRepository.findById(uuid).orElseThrow(InvalidDepartmentIdException::new);
-            return DepartmentMapper.INSTANCE.departmentToDepartmentDTO(department);
-        } catch (IllegalArgumentException e) {
-            throw new IllegalArgumentException(e.getMessage());
+            return true;
+        } catch (IllegalArgumentException ex) {
+            return false;
         }
     }
 
@@ -69,16 +78,24 @@ public class DepartmentImpl implements DepartmentService {
         if (department != null) {
             return DepartmentMapper.INSTANCE.departmentToDepartmentDTO(department);
         } else {
-            throw new IllegalArgumentException("Department not found");
+            throw new BadRequestException("No such department with the name");
         }
     }
 
 
     @Override
     public DepartmentDTO deleteDepartmentById(String id) {
-        Department department = departmentRepository.findById(UUID.fromString(id)).orElseThrow(InvalidDepartmentIdException::new);
-        departmentRepository.delete(department);
-        return DepartmentMapper.INSTANCE.departmentToDepartmentDTO(department);
+        // kako da frli vakov exception
+        if(id==null || id.isEmpty()) {
+            throw new BadRequestException("No id was passed");
+        }
+        if(isValidUUID(id)) {
+            Department department = departmentRepository.findById(UUID.fromString(id)).orElseThrow(InvalidDepartmentIdException::new);
+            departmentRepository.delete(department);
+            return DepartmentMapper.INSTANCE.departmentToDepartmentDTO(department);
+        } else {
+            throw new BadRequestException("Invalid id");
+        }
     }
 
     @Override
@@ -87,7 +104,7 @@ public class DepartmentImpl implements DepartmentService {
         if (department != null) {
             departmentRepository.delete(department);
         } else {
-            throw new IllegalArgumentException("Department not found");
+            throw new DepartmentNotFoundException("Department not found");
         }
         return DepartmentMapper.INSTANCE.departmentToDepartmentDTO(department);
     }
